% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralEvaluation.r
\name{f1_records}
\alias{f1_records}
\title{Evaluate the performance of classification schemes}
\usage{
f1_records(results, db, vr, clsfr)
}
\arguments{
\item{results}{The results of the classification scheme}

\item{db}{The database used}

\item{vr}{The identifier used}

\item{clsfr}{The classifier used}
}
\value{
A very long list of dataframes.

Dataframe column dictionary
\itemize{
\item true_match (numeric): Number of true matches (TM) in the dataset
\item false_match (numeric): Number of false matches (FM) in the dataset
\item cellC (numeric): The number of missed matches (MM)
\item recall (numeric): Recall is calculated by TM/(MM + TM)
\item precision (numeric): Precision is calculated by TM/(FM + TM)
\item fmeasure (numeric): F1 is calculated by (weight * R) + ((1-weight) * P)
\item weight (numeric): F1 can be weighted toward either Precision or Recall, but is equally weighted in this study
\item weight_r (numeric): Intended to use, but wound up ignoring
\item weight_p (numeric): Intended to use, but wound up ignoring
\item database (character): Database used
\item classifier (character): Classifier used
\item region (character): Identifier used
\item confidence (numeric): The number of times the same identification occured under random sampling (bootstrapping)
}
}
\description{
A very long and convoluted function that I'll most
likely split into smaller chunks at a later time. Meant to be called from the function \code{total_results()}

During development, I got a little lazy with the column names after the python script
is done reformatting classification results. The classifer column
is labled 'blca', whether or not the actual classifier used was BLCA
or Naive Bayes. I probably will not change that.

This function also includes lists of bacteria species that are missing
from the various databases used in the study. Until I set up these
lists as imported data, they live here.
}
