% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralEvaluation.r
\name{list_of_df}
\alias{list_of_df}
\title{Make a list of dataframes}
\usage{
list_of_df(path_to_dir, prefix)
}
\arguments{
\item{path_to_dir}{The path to the directory containing the output from the classification schemes.}

\item{prefix}{Which classifier was used, e.g. BLCA, Naive Bayes}
}
\value{
A list of dataframes
}
\description{
I got tired of loading all the csv files individually,
so this function does it in batch. Reads all filenames in a directory,
then creates a dataframe from the \code{csv} files, then adds names to the growing list of dataframes.
}
